Convert position and size properties into decimal type.

+++ src/ShapeCrawler/Presentations/SlideSize.cs
@@ -13,17 +13,17 @@ internal sealed class SlideSize
         this.pSlideSize = pSlideSize;
     }
 
-    internal int Width() => UnitConverter.HorizontalEmuToPixel(this.pSlideSize.Cx!.Value);
+    internal decimal Width() => UnitConverter.HorizontalEmuToPixel(this.pSlideSize.Cx!.Value);
 
-    internal int Height() => UnitConverter.HorizontalEmuToPixel(this.pSlideSize.Cy!.Value);
+    internal decimal Height() => UnitConverter.HorizontalEmuToPixel(this.pSlideSize.Cy!.Value);
 
-    internal void UpdateWidth(int pixels)
+    internal void UpdateWidth(decimal pixels)
     {
         var emu = UnitConverter.HorizontalPixelToEmu(pixels);
         this.pSlideSize.Cx = new Int32Value((int)emu);
     }
     
-    internal void UpdateHeight(int pixels)
+    internal void UpdateHeight(decimal pixels)
     {
         var emu = UnitConverter.VerticalPixelToEmu(pixels);
         this.pSlideSize.Cy = new Int32Value((int)emu);
